Fix bug in evicting cached compiled expressions
Added a unittest for this as well.  This also included
added a default equality comparison to the AST nodes.

+++ jmespath/parser.py
@@ -81,7 +81,7 @@ class Parser(object):
     def _free_cache_entries(self):
         # This logic is borrowed from the new regex library which
         # uses similar eviction strategies.
-        for key in random.sample(self._cache.keys(), self._max_size / 2):
+        for key in random.sample(self._cache.keys(), int(self._max_size / 2)):
             del self._cache[key]
 
     @classmethod
